# Conflicts:
#	lib/Src/pages/login/login.dart

+++ lib/Src/pages/heroAltar/heroAltarClock.dart
@@ -2,32 +2,35 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:upgradegame/Common/app/config.dart';
 
-class HeroAltarClock extends StatefulWidget{
-
+class HeroAltarClock extends StatefulWidget {
   String imageUrl;
-  int adIconHeight;
+  double adIconHeight;
   List<int> remainDays;
 
-  HeroAltarClock({Key key,this.remainDays,this.adIconHeight,this.imageUrl}):super(key:key);
+  HeroAltarClock({Key key, this.remainDays, this.adIconHeight, this.imageUrl})
+      : super(key: key);
+
   @override
   _HeroAltarClockState createState() => _HeroAltarClockState();
-
 }
 
-class _HeroAltarClockState extends State<HeroAltarClock>{
-
-  Widget buildList(){
-    List <Widget> clockDayList = [];
+class _HeroAltarClockState extends State<HeroAltarClock> {
+  Widget buildList() {
+    List<Widget> clockDayList = [];
     Widget content;
-    for(int i=0;i<this.widget.remainDays.length;i++){
+    for (int i = 0; i < this.widget.remainDays.length; i++) {
       clockDayList.add(
-          Row(
-            children: <Widget>[
-              Image(image: new AssetImage(this.widget.imageUrl), height:ScreenUtil().setHeight(this.widget.adIconHeight)),
-              Text(this.widget.remainDays[i].toString()+'天',style: CustomFontSize.textStyle16,),
-            ],
-          ),
-
+        Row(
+          children: <Widget>[
+            Image(
+                image: new AssetImage(this.widget.imageUrl),
+                height: ScreenUtil().setHeight(this.widget.adIconHeight)),
+            Text(
+              this.widget.remainDays[i].toString() + '天',
+              style: CustomFontSize.textStyle16,
+            ),
+          ],
+        ),
       );
     }
     content = new Row(
